:truck: Refactored into package. Separated models from bot.
