--- slack_bot/__init__.py
@@ -1,1 +1,0 @@
-from .strategies import *
