+++ slack_bot/slack_bot/__init__.py
@@ -1,0 +1,2 @@
+from .bot import Bot
+from .models import MODELS
