+++ slack_bot/slack_bot/bot/__init__.py
@@ -1,0 +1,3 @@
+from .bot import Bot
+
+__all__ = ["Bot"]
