+++ slack_bot/slack_bot/models/__init__.py
@@ -1,0 +1,8 @@
+from .base import ResponseModel
+from .hello import Hello
+
+MODELS = {
+    "hello": Hello,
+}
+
+__all__ = ["MODELS", "ResponseModel"]
