--- slack_bot/strategies/__init__.py
@@ -1,1 +1,0 @@
-from .hello import Hello
